feat: Add navigation in DbManager
