+++ USca/DbManager/MainWindow.xaml.cs
@@ -1,14 +1,6 @@
-﻿using USca_DbManager.User;
-using RestSharp;
-using System;
-using System.Diagnostics;
-using System.Net.Http;
-using System.Text.Json;
-using System.Threading;
-using System.Windows;
-using System.Windows.Input;
-using USca_DbManager.Tags;
-using System.Linq;
+﻿using System.Windows;
+using USca_DbManager.User;
+using USca_DbManager.Util;
 
 namespace USca_DbManager
 {
@@ -17,39 +9,12 @@ namespace USca_DbManager
 		public MainWindow()
 		{
 			InitializeComponent();
+			Navigate(new Login());
 		}
 
-		private void btnLogin_Click(object sender, RoutedEventArgs e)
+		public void Navigate(MyPage page)
 		{
-			string username = tbUsername.Text;
-			string password = tbPassword.Password;
-			DoLogin(username, password);
-		}
-
-		private async void DoLogin(string username, string password)
-		{
-			btnLogin.IsEnabled = false;
-			Mouse.OverrideCursor = Cursors.Wait;
-
-			try
-			{
-				var res = await UserService.Login(username, password);
-				MessageBox.Show($"Logged in as {res.username}", "Alert.", MessageBoxButton.OK, MessageBoxImage.Exclamation);
-
-				var res2 = await TagService.GetAllTags();
-				Console.WriteLine(res2);
-			}
-			catch (BadCredentialsException)
-			{
-				MessageBox.Show("Username or Password invalid", "Could not log in.", MessageBoxButton.OK, MessageBoxImage.Error);
-				tbPassword.Clear();
-				tbPassword.Focus();
-			}
-			finally
-			{
-				Mouse.OverrideCursor = null;
-				btnLogin.IsEnabled = true;
-			}
+			FrameContent.Navigate(page);
 		}
 	}
 }
