+++ USca/DbManager/Tags/TagsDashboard.xaml.cs
@@ -1,0 +1,12 @@
+﻿using USca_DbManager.Util;
+
+namespace USca_DbManager.Tags
+{
+	public partial class TagsDashboard : MyPage
+	{
+		public TagsDashboard()
+		{
+			InitializeComponent();
+		}
+	}
+}
