+++ USca/DbManager/Util/WinPage.cs
@@ -1,0 +1,10 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+
+namespace USca_DbManager.Util
+{
+	public partial class MyPage : Page
+	{
+		public MainWindow Owner { get { return (MainWindow)Window.GetWindow(this); } }
+	}
+}
